+++ spec/stack_master/template_compilers/yaml_erb_spec.rb
@@ -4,18 +4,28 @@ RSpec.describe StackMaster::TemplateCompilers::YamlErb do
   before(:all) { described_class.require_dependencies }
 
   describe '.compile' do
-    let(:compile_time_parameters) { { 'SubnetCidrs' => ['10.0.0.0/28:ap-southeast-2', '10.0.2.0/28:ap-southeast-1'] } }
-
-    def compile
-      described_class.compile(stack_definition.template_dir, stack_definition.template, compile_time_parameters)
+    subject(:compile) do
+      described_class.compile(
+        stack_definition.template_dir,
+        stack_definition.template,
+        compile_time_parameters
+      )
     end
 
     context 'a YAML template using a loop over compile time parameters' do
-      let(:stack_definition) { StackMaster::StackDefinition.new(template_dir: 'spec/fixtures/templates/erb',
-                                                                template: 'compile_time_parameters_loop.yml.erb') }
+      let(:stack_definition) do
+        StackMaster::StackDefinition.new(
+          template_dir: 'spec/fixtures/templates/erb',
+          template: 'compile_time_parameters_loop.yml.erb'
+        )
+      end
+
+      let(:compile_time_parameters) do
+        { 'SubnetCidrs' => ['10.0.0.0/28:ap-southeast-2', '10.0.2.0/28:ap-southeast-1'] }
+      end
 
       it 'renders the expected output' do
-        expect(compile).to eq <<~EOEXPECTED
+        expect(compile).to eq(<<~YAML)
           ---
           Description: "A test case for generating subnet resources in a loop"
           Parameters:
@@ -39,7 +49,47 @@ RSpec.describe StackMaster::TemplateCompilers::YamlErb do
                 VpcId: !Ref Vpc
                 CidrBlock: 10.0.2.0/28
                 AvailabilityZone: ap-southeast-1
-          EOEXPECTED
+          YAML
+      end
+    end
+
+    context 'a YAML template using loading a userdata script from an external file' do
+      let(:stack_definition) do
+        StackMaster::StackDefinition.new(
+          template_dir: 'spec/fixtures/templates/erb',
+          template: 'user_data.yml.erb'
+        )
+      end
+
+      let(:compile_time_parameters) { {} }
+
+      it 'renders the expected output' do
+        expect(compile).to eq(<<~YAML)
+          Description: A test case for storing the userdata script in a dedicated file
+
+          Resources:
+            LaunchConfig:
+              Type: 'AWS::AutoScaling::LaunchConfiguration'
+              Properties:
+                UserData: {
+            "Fn::Base64": {
+              "Fn::Join": [
+                "",
+                [
+                  "#!/bin/bash\\n",
+                  "\\n",
+                  "echo 'Hello, World!'\\n",
+                  "REGION=",
+                  {
+                    "Ref": "AWS::Region"
+                  },
+                  "\\n",
+                  "echo $REGION\\n"
+                ]
+              ]
+            }
+          }
+        YAML
       end
     end
   end